% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copy_icons}
\alias{copy_icons}
\title{Copy icons to directory}
\usage{
copy_icons(path)
}
\arguments{
\item{path}{Ruta del directorio destino.}
}
\value{
Invisible, copia los archivos de iconos.
}
\description{
Copia los iconos del paquete a un directorio especificado.
}
