% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_charts.R
\name{create_bar_chart}
\alias{create_bar_chart}
\title{Create Bar Chart}
\usage{
create_bar_chart(data, options)
}
\arguments{
\item{data}{Prepared chart data}

\item{options}{Chart options}
}
\value{
hgmagic bar chart
}
\description{
Creates a bar chart using hgmagic
}
