% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_charts.R
\name{create_donut_chart}
\alias{create_donut_chart}
\title{Create Donut Chart}
\usage{
create_donut_chart(data, options)
}
\arguments{
\item{data}{Prepared chart data}

\item{options}{Chart options}
}
\value{
hgmagic donut chart
}
\description{
Creates a donut chart using hgmagic
}
