% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_charts.R
\name{create_pie_chart}
\alias{create_pie_chart}
\title{Create Pie Chart}
\usage{
create_pie_chart(data, options)
}
\arguments{
\item{data}{Prepared chart data}

\item{options}{Chart options}
}
\value{
hgmagic pie chart
}
\description{
Creates a pie chart using hgmagic
}
