% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_charts.R
\name{create_treemap_chart}
\alias{create_treemap_chart}
\title{Create Treemap Chart}
\usage{
create_treemap_chart(data, options)
}
\arguments{
\item{data}{Prepared chart data}

\item{options}{Chart options}
}
\value{
hgmagic treemap chart
}
\description{
Creates a treemap chart using hgmagic
}
