% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_download.R
\name{downloadTableServer}
\alias{downloadTableServer}
\title{Download Table Server}
\usage{
downloadTableServer(
  id,
  element = NULL,
  formats,
  file_prefix = "table",
  debug = FALSE
)
}
\arguments{
\item{id}{Module ID}

\item{element}{Reactive expression returning data to download}

\item{formats}{File formats to support}

\item{file_prefix}{Prefix for downloaded files}

\item{debug}{Boolean to control console debug output}
}
\description{
Download Table Server
}
