% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_download.R
\name{downloadTableUI}
\alias{downloadTableUI}
\title{Download Table UI}
\usage{
downloadTableUI(
  id,
  text = "Download",
  formats = NULL,
  display = c("buttons", "dropdown"),
  dropdownLabel = "Download",
  dropdownWidth = 150
)
}
\arguments{
\item{id}{Module ID}

\item{text}{Download button text}

\item{formats}{File formats to support}

\item{display}{Display type: "buttons" or "dropdown"}

\item{dropdownLabel}{Label for dropdown}

\item{dropdownWidth}{Width of dropdown}
}
\description{
Download Table UI
}
