% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotable_fun.R
\name{duckdb_con}
\alias{duckdb_con}
\title{Create DuckDB connection}
\usage{
duckdb_con(dbdir, read_only = TRUE)
}
\arguments{
\item{dbdir}{Path to DuckDB database file.}

\item{read_only}{Logical, indicates if the connection is read-only (defaults
to \code{TRUE}).}
}
\value{
DuckDB connection object.
}
\description{
Create DuckDB connection
}
\keyword{internal}
