% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotable_fun.R
\name{duckdb_disconnect}
\alias{duckdb_disconnect}
\title{Disconnect DuckDB connection}
\usage{
duckdb_disconnect(con)
}
\arguments{
\item{con}{DuckDB connection to disconnect.}
}
\value{
\code{NULL} (invisible).
}
\description{
Disconnect DuckDB connection
}
