% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotable_fun.R
\name{duckdb_list_tables}
\alias{duckdb_list_tables}
\title{List tables in DuckDB connection}
\usage{
duckdb_list_tables(con)
}
\arguments{
\item{con}{DuckDB connection.}
}
\value{
Character vector of table names.
}
\description{
List tables in DuckDB connection
}
