% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotable_fun.R
\name{duckdb_load_ext}
\alias{duckdb_load_ext}
\title{Load DuckDB extension}
\usage{
duckdb_load_ext(ext_name, con)
}
\arguments{
\item{ext_name}{Name of the extension to load (e.g., "spatial").}

\item{con}{DuckDB connection.}
}
\value{
Invisible, loads the extension.
}
\description{
Load DuckDB extension
}
\keyword{internal}
