% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotable_fun.R
\name{duckdb_read_geotable}
\alias{duckdb_read_geotable}
\title{Read geotable from DuckDB}
\usage{
duckdb_read_geotable(tblname, con = NULL, geometrycol = "geom")
}
\arguments{
\item{tblname}{Table name.}

\item{con}{DuckDB connection (optional, defaults to \code{NULL}).}

\item{geometrycol}{Name of geometry column (defaults to "geom").}
}
\value{
\code{sf} object.
}
\description{
Read geotable from DuckDB
}
