% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotable_fun.R
\name{duckdb_read_table}
\alias{duckdb_read_table}
\title{Read table from DuckDB}
\usage{
duckdb_read_table(tblname, con = NULL, collect = FALSE)
}
\arguments{
\item{tblname}{Table name.}

\item{con}{DuckDB connection (optional, defaults to \code{NULL}).}

\item{collect}{Logical, whether to collect the result (defaults to \code{FALSE}).}
}
\value{
Table object (tbl or data frame).
}
\description{
Read table from DuckDB
}
\keyword{internal}
