% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimadas.R
\name{estimadas_grupo}
\alias{estimadas_grupo}
\title{Get estimated species by grupo}
\usage{
estimadas_grupo(grupo, con)
}
\arguments{
\item{grupo}{Slug del grupo biológico.}

\item{con}{Conexión a la base de datos.}
}
\value{
Lista con estimaciones de especies.
}
\description{
Obtiene estimaciones de especies amenazadas, CITES, endémicas para un grupo
biológico.
}
