% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_chart_selector.R
\name{exp_chart_selector_ui}
\alias{exp_chart_selector_ui}
\title{Chart Selector UI Module}
\usage{
exp_chart_selector_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
UI elements for chart selector
}
\description{
Creates an image-based chart selector using shinyinvoer::buttonImageInput
with dynamic chart availability based on data combinations
}
