% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_debug2.R
\name{exp_debug2_server}
\alias{exp_debug2_server}
\title{Debug Module 2 Server - Visualization Debug}
\usage{
exp_debug2_server(id, r, debug = FALSE)
}
\arguments{
\item{id}{Module ID}

\item{r}{Reactive values object}

\item{debug}{Boolean to control whether debug output is shown}
}
\description{
Debug Module 2 Server - Visualization Debug
}
