% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_debug2.R
\name{exp_debug2_ui}
\alias{exp_debug2_ui}
\title{Debug Module 2 UI - Visualization Debug}
\usage{
exp_debug2_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\description{
Debug Module 2 UI - Visualization Debug
}
