% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_debug.R
\name{exp_debug_server}
\alias{exp_debug_server}
\title{Debug Module Server}
\usage{
exp_debug_server(id, r, debug = FALSE)
}
\arguments{
\item{id}{Module ID}

\item{r}{Reactive values object}

\item{debug}{Boolean to control whether debug output is shown}
}
\description{
Debug Module Server
}
