% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_debug.R
\name{exp_debug_ui}
\alias{exp_debug_ui}
\title{Debug Module UI}
\usage{
exp_debug_ui(id)
}
\arguments{
\item{id}{Module ID}

\item{debug}{Boolean to control whether debug output is shown}
}
\description{
Debug Module UI
}
