% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_inputs_grupo.R
\name{exp_inputs_grupo_ui}
\alias{exp_inputs_grupo_ui}
\title{Grupo Type Selection UI Module}
\usage{
exp_inputs_grupo_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
UI elements for grupo type selection
}
\description{
Creates a hierarchical checkbox interface for grupo type selection
}
