% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_inputs_tematica.R
\name{exp_inputs_tematica_server}
\alias{exp_inputs_tematica_server}
\title{Thematic Selection Server Module}
\usage{
exp_inputs_tematica_server(id, con, session_main = NULL, debug = FALSE)
}
\arguments{
\item{id}{Module ID}

\item{session_main}{Main session object (optional)}

\item{debug}{Boolean to control console debug output}

\item{tematicas_tree}{Thematic tree structure from database}
}
\value{
Reactive expression returning the selected tematica slug
}
\description{
Handles the server logic for hierarchical thematic selection
}
