% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_inputs_tematica.R
\name{exp_inputs_tematica_ui}
\alias{exp_inputs_tematica_ui}
\title{Thematic Selection UI Module}
\usage{
exp_inputs_tematica_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
UI elements for thematic selection
}
\description{
Creates a hierarchical checkbox interface for thematic selection
}
