% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_inputs.R
\name{exp_inputs_ui}
\alias{exp_inputs_ui}
\title{Input Controls UI Module}
\usage{
exp_inputs_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
UI elements for input controls
}
\description{
Creates the left panel input controls for region, group, and theme selection
}
