% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_species_table.R
\name{exp_species_table_ui}
\alias{exp_species_table_ui}
\title{Species Table UI Module}
\usage{
exp_species_table_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
UI elements for species table
}
\description{
Creates the right panel species table with summary and download functionality
}
