% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_visualization3.R
\name{exp_visualization3_ui}
\alias{exp_visualization3_ui}
\title{Visualization UI Module}
\usage{
exp_visualization3_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
UI elements for visualization
}
\description{
Creates the center panel visualization with chart selector and dynamic controls
}
