% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_visualization4_map.R
\name{exp_visualization4_map_server}
\alias{exp_visualization4_map_server}
\title{Map Visualization Server Module}
\usage{
exp_visualization4_map_server(output, r, con, debug = FALSE)
}
\arguments{
\item{output}{The output object from parent module}

\item{r}{Reactive values object}

\item{con}{Database connection}

\item{debug}{Boolean to control console debug output}
}
\description{
Renders leaflet map based on reactive values
This is called as a nested module within exp_visualization4
}
