% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_visualization4.R
\name{exp_visualization4_server}
\alias{exp_visualization4_server}
\title{Visualization Server Module}
\usage{
exp_visualization4_server(id, r, con, loading_fns = NULL, debug = FALSE)
}
\arguments{
\item{id}{Module ID}

\item{r}{Reactive values object}

\item{con}{Database connection}

\item{loading_fns}{Optional list with show/hide functions for loading overlay}

\item{debug}{Boolean to control console debug output}
}
\description{
Handles the server logic for visualization including chart rendering and data modals
All input controls, chart selector, tipo, breadcrumb, and indicador handled by exp_viz_inputs
}
