% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_visualization4.R
\name{exp_visualization4_ui}
\alias{exp_visualization4_ui}
\title{Visualization UI Module}
\usage{
exp_visualization4_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
UI elements for visualization
}
\description{
Creates the center panel visualization - download buttons and viz area only
}
