% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_visualization.R
\name{exp_visualization_ui}
\alias{exp_visualization_ui}
\title{Visualization UI Module}
\usage{
exp_visualization_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
UI elements for visualization
}
\description{
Creates the center panel visualization with chart selector and dynamic controls
}
