% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_viz_inputs.R
\name{exp_viz_inputs_server}
\alias{exp_viz_inputs_server}
\title{Visualization Inputs Server Module}
\usage{
exp_viz_inputs_server(id, r, debug = FALSE)
}
\arguments{
\item{id}{Module ID}

\item{r}{Reactive values object (shared)}

\item{debug}{Boolean to control console debug output}
}
\description{
Syncs tipo and chart type with reactive values, computes tematica,
indicador, and breadcrumb, and manages chart availability.
}
