% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_viz_inputs.R
\name{exp_viz_inputs_ui}
\alias{exp_viz_inputs_ui}
\title{Visualization Inputs UI Module}
\usage{
exp_viz_inputs_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\description{
Provides controls for visualization type (tipo) and chart selector,
and displays breadcrumb.
}
