% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_utils_ui.R
\name{format_species_data}
\alias{format_species_data}
\title{Format species data for display}
\usage{
format_species_data(data)
}
\arguments{
\item{data}{Raw species data}
}
\value{
Formatted data frame
}
\description{
Format species data for display
}
