% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotable_fun.R
\name{geotable_functions}
\alias{geotable_functions}
\title{Functions from the geotable package used in sibdata}
\description{
This file contains functions from the \code{geotable} package that are used in
the \code{sibdata} code, transcribed so they can be used without directly
accessing the geotable package.
}
\keyword{internal}
