% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_utils_ui.R
\name{get_app_connection}
\alias{get_app_connection}
\title{Database connection helper}
\usage{
get_app_connection(path = NULL, debug = FALSE)
}
\arguments{
\item{path}{Path to database file (optional)}

\item{debug}{Boolean to control console debug output}
}
\value{
DBI connection to SQLite database
}
\description{
Database connection helper
}
