% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_utils_ui.R
\name{get_app_options}
\alias{get_app_options}
\title{Get available options for dropdowns}
\usage{
get_app_options(con, debug = FALSE)
}
\arguments{
\item{con}{Database connection}

\item{debug}{Boolean to control console debug output}
}
\value{
List of options for UI inputs
}
\description{
Get available options for dropdowns
}
