% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_chart_selector.R
\name{get_available_charts}
\alias{get_available_charts}
\title{Get Chart Availability Logic}
\usage{
get_available_charts(tipo, tematica)
}
\arguments{
\item{tipo}{Data type ("registros" or "especies")}

\item{tematica}{Thematic category}
}
\value{
Named vector of available charts
}
\description{
Determines which charts are available based on current data combination
This replicates the logic from the original app (lines 340-355)
}
