% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_charts.R
\name{get_chart_function_name}
\alias{get_chart_function_name}
\title{Get Chart Function Name}
\usage{
get_chart_function_name(chart_type)
}
\arguments{
\item{chart_type}{Type of chart}
}
\value{
Function name for hgmagic
}
\description{
Maps chart types to their corresponding hgmagic function names
Based on original app logic (line 819)
}
