% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_charts.R
\name{get_chart_options}
\alias{get_chart_options}
\title{Get Chart Options}
\usage{
get_chart_options(chart_type, r)
}
\arguments{
\item{chart_type}{Type of chart}

\item{r}{Reactive values object}
}
\value{
List of chart options
}
\description{
Generates chart options including palette and styling
Based on original app logic (lines 734-787)
}
