% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_utils_ui.R
\name{get_species_table_options}
\alias{get_species_table_options}
\title{Create DT options with custom styling}
\usage{
get_species_table_options()
}
\value{
List of DT options
}
\description{
Create DT options with custom styling
}
