% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_data_controls.R
\name{get_subcategory_display_name}
\alias{get_subcategory_display_name}
\title{Get Subcategory Display Name}
\usage{
get_subcategory_display_name(subcategory, theme)
}
\arguments{
\item{subcategory}{Subcategory value}

\item{theme}{Theme type ("amenazadas", "cites", "exoticas")}
}
\value{
Display name for subcategory
}
\description{
Gets the display name for a subcategory value
}
