% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_utils_ui.R, R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Null-coalescing operator}
\usage{
x \%||\% y

x \%||\% y
}
\arguments{
\item{x}{Valor a verificar.}

\item{y}{Valor por defecto.}
}
\value{
\code{x} si no es NULL/NA/vacío, de lo contrario \code{y}.
}
\description{
Operador que retorna el segundo argumento si el primero es NULL, NA o vacío.
}
\keyword{internal}
