% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotable_fun.R
\name{gt_available_main_maps}
\alias{gt_available_main_maps}
\title{Get available main maps}
\usage{
gt_available_main_maps(con = NULL)
}
\arguments{
\item{con}{Connection (optional, defaults to \code{NULL}).}
}
\value{
Character vector of main map names.
}
\description{
Get available main maps
}
