% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotable_fun.R
\name{gt_con}
\alias{gt_con}
\title{Connection to the geotable database}
\usage{
gt_con(con = NULL, read_only = TRUE)
}
\arguments{
\item{con}{Existing connection (optional, defaults to \code{NULL}).}

\item{read_only}{Logical, indicates if the connection is read-only (defaults
to \code{TRUE}).}
}
\value{
Connection object to the geotable database.
}
\description{
Connection to the geotable database
}
