% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotable_fun.R
\name{gt_match}
\alias{gt_match}
\title{Match geotable Data to a geotable Map}
\usage{
gt_match(
  d,
  map_name,
  col = NULL,
  centroids = TRUE,
  unique = FALSE,
  con = NULL,
  threshold = 0.8,
  strict_threshold = TRUE
)
}
\arguments{
\item{d}{The input data frame to be matched.}

\item{map_name}{The name of the geotable map.}

\item{col}{The column name or index to be used for matching. If NULL, the function
will try to guess the appropriate column for matching.}

\item{centroids}{A logical value indicating whether to include centroids in the
output. Defaults to TRUE.}
}
\value{
A data frame with the matched data and centroids if requested with ..gd_id and ..gd_name.
}
\description{
This function matches input data to a geotable map, based on the provided map_name
and the specified column.
}
\examples{
data <- readr::read_csv("data-raw/sample_data/col_departments/plebiscito_2016_departamentos.csv")
gd_match(data, map_name = "col_departments")
}
