% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotable_fun.R
\name{gt_match_names}
\alias{gt_match_names}
\title{Match geotable Data to a geotable Map by Names}
\usage{
gt_match_names(
  d,
  map_name = NULL,
  col = NULL,
  codes = NULL,
  altnames = NULL,
  con = NULL
)
}
\arguments{
\item{d}{The input data frame to be matched.}

\item{map_name}{The name of the geotable map. Defaults to NULL.}

\item{col}{Optional column with name to be used for matching by names. Defaults to NULL.}

\item{codes}{Optional codes data frame to be used for matching. Defaults to NULL.}

\item{altnames}{Optional alternative names data frame to be used for matching. Defaults to NULL.}
}
\value{
A data frame with the matched data based on names.
}
\description{
This function matches input data to a geotable map based on the provided map_name
and the specified name column.
}
\examples{
data <- data.frame(id_name = c("Antioquia", "Bogota", "Quindio", "Caldas"), value = runif(4))
gd_match_names(data, map_name = "col_departments")
}
