% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotable_fun.R
\name{gt_regions}
\alias{gt_regions}
\title{Get regions for a map}
\usage{
gt_regions(map_name = NULL, con)
}
\arguments{
\item{map_name}{Map name (optional, defaults to \code{NULL}).}

\item{con}{Connection.}
}
\value{
Table with region information.
}
\description{
Get regions for a map
}
