% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotable_fun.R
\name{gt_valid_map_name}
\alias{gt_valid_map_name}
\title{Validate map name}
\usage{
gt_valid_map_name(map_name, con = NULL)
}
\arguments{
\item{map_name}{Map name to validate.}

\item{con}{Connection (optional, defaults to \code{NULL}).}
}
\value{
Invisible, throws error if map name is not valid.
}
\description{
Validate map name
}
\keyword{internal}
