% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotable_fun.R
\name{gt_which_geoname_col}
\alias{gt_which_geoname_col}
\title{Identify the column with geoname data in a dataframe}
\usage{
gt_which_geoname_col(
  d,
  map_name,
  con = NULL,
  threshold = 0.5,
  strict_threshold = TRUE
)
}
\arguments{
\item{d}{A dataframe containing the data to be analyzed.}

\item{map_name}{A character string specifying the name of the map to be used,
you can view available maps with geotable::available_maps().}
}
\value{
The name of the column with geoname data, or NA if no column meets the
90\\% agreement threshold.
}
\description{
This function takes a dataframe and a map name as inputs, and returns the name
of the column that has the geoname data for the specified map. If data have
more than 50 rows, the function looks at the first 50 rows of the dataframe
and checks if any column has values that match the geocode ids for the given map.
The column with the highest number of matching geocode ids is returned, provided
that at least a user-specified percentage of the values in that column agree
with the geoname.
}
\examples{
df <- data.frame(country = c("Argentina", "Colombia", "Angora", "Brasil"), value = runif(4))
which_geoname_col(df, "world_countries")
}
