% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_pages.R
\name{info_publicador}
\alias{info_publicador}
\title{Generate publicador information}
\usage{
info_publicador(con)
}
\arguments{
\item{con}{Conexión a la base de datos.}
}
\value{
Lista con información de publicadores por región y filtros de
navegación.
}
\description{
Genera información sobre publicadores de datos de biodiversidad, incluyendo
estadísticas por región y filtros de navegación.
}
