% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.empty}
\alias{is.empty}
\title{Check if object is empty}
\usage{
is.empty(x)
}
\arguments{
\item{x}{Objeto a verificar.}
}
\value{
Logical indicando si el objeto está vacío.
}
\description{
Check if object is empty
}
\keyword{internal}
