% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotable_fun.R
\name{is_main_map}
\alias{is_main_map}
\title{Check if map is a main map}
\usage{
is_main_map(map_name)
}
\arguments{
\item{map_name}{Map name.}
}
\value{
Logical indicating if it's a main map.
}
\description{
Check if map is a main map
}
\keyword{internal}
