% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_especies.R
\name{list_species}
\alias{list_species}
\title{List species for a region}
\usage{
list_species(
  region,
  grupo = NULL,
  tematica = NULL,
  with_labels = FALSE,
  con = NULL
)
}
\arguments{
\item{region}{Slug de la región.}

\item{grupo}{Slug del grupo biológico (opcional).}

\item{tematica}{Slug de la temática (opcional).}

\item{with_labels}{Logical, retornar con etiquetas en español (default:
\code{FALSE}).}

\item{con}{Conexión a la base de datos.}
}
\value{
Objeto \code{tbl} o data frame con información de especies.
}
\description{
Obtiene lista de especies para una región, opcionalmente filtradas por grupo
o temática.
}
