% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeup_fun.R
\name{makeup_chr}
\alias{makeup_chr}
\title{Format character strings}
\usage{
makeup_chr(x, style = "Title")
}
\arguments{
\item{x}{Vector de caracteres a formatear.}

\item{style}{Estilo de formato. Opciones: "Title", "UPPER", "lower", etc.}
}
\value{
Vector de caracteres formateado.
}
\description{
Formatea cadenas de caracteres según el estilo especificado.
}
\examples{
makeup_chr("hello world", "Title")
#> [1] "Hello World"

}
