% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parent.R
\name{parent_tematica}
\alias{parent_tematica}
\title{Get parent region tematica data}
\usage{
parent_tematica(region)
}
\arguments{
\item{region}{Slug de la región.}
}
\value{
Objeto \code{tbl} con datos de temática de la región padre.
}
\description{
Obtiene datos de temática de la región padre.
}
\keyword{internal}
