% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_col}
\alias{parse_col}
\title{Parse column name}
\usage{
parse_col(d, col = NULL)
}
\arguments{
\item{d}{Data frame.}

\item{col}{Nombre o índice de columna (opcional, por defecto usa la primera
columna).}
}
\value{
Nombre de la columna.
}
\description{
Convierte un índice numérico o nombre de columna en el nombre de columna
válido.
}
