% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{paste_dash}
\alias{paste_dash}
\title{Paste string with dashes}
\usage{
paste_dash(str, times = 1)
}
\arguments{
\item{str}{Cadena de texto.}

\item{times}{Número de niveles de indentación (guiones).}
}
\value{
Cadena formateada con guiones.
}
\description{
Crea una cadena con guiones para formatear texto jerárquico.
}
