% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkg_resources}
\alias{pkg_resources}
\title{Get package resources path}
\usage{
pkg_resources(...)
}
\arguments{
\item{...}{Componentes de la ruta del recurso.}
}
\value{
Ruta completa del recurso del paquete.
}
\description{
Get package resources path
}
\keyword{internal}
