% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgmagic_utils.R
\name{process_CatNum}
\alias{process_CatNum}
\title{Data processing for visualization}
\usage{
process_CatNum(d, viz)
}
\arguments{
\item{d}{A data frame containing the data to process. The structure of the data frame varies depending on the processing function.}

\item{viz}{The desired type of visualisation. It can be "bar", "column", "pie", or "donut".}
}
\value{
A list with the processed data for visualisation.
}
\description{
This set of functions provides tools to process data for visualisation in different types of charts.
}
\examples{
# Example usage of data processing functions
d <- data.frame(
  category = c("A", "A", "B", "B"),
  value = c(10, 20, 30, 40)
)
process_CatNum(d, "bar")
}
