% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{region_grupo}
\alias{region_grupo}
\title{Get region grupo data}
\usage{
region_grupo(region, grupo, con)
}
\arguments{
\item{region}{Slug de la región.}

\item{grupo}{Slug del grupo biológico.}

\item{con}{Conexión a la base de datos.}
}
\value{
Objeto \code{tbl} con datos de grupo de la región.
}
\description{
Obtiene datos de grupo biológico para una región.
}
\keyword{internal}
