% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_indicadores.R
\name{region_indicadores}
\alias{region_indicadores}
\title{Get region indicators}
\usage{
region_indicadores(region, indicadores, con = NULL)
}
\arguments{
\item{region}{Slug de la región.}

\item{indicadores}{Vector con nombres de indicadores.}

\item{con}{Conexión a la base de datos.}
}
\value{
Lista con valores de indicadores.
}
\description{
Obtiene valores de indicadores específicos para una región.
}
