% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{region_tematica}
\alias{region_tematica}
\title{Get region tematica data}
\usage{
region_tematica(region, con)
}
\arguments{
\item{region}{Slug de la región.}

\item{con}{Conexión a la base de datos.}
}
\value{
Objeto \code{tbl} con datos de temática de la región.
}
\description{
Obtiene datos de temática para una región.
}
\keyword{internal}
