% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run Shiny application}
\usage{
run_app()
}
\value{
Invisible, ejecuta la aplicación Shiny en el puerto 3838.
}
\description{
Ejecuta la aplicación Shiny del paquete sibdata.
}
