% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_download.R
\name{saveTable}
\alias{saveTable}
\title{Save Table to File}
\usage{
saveTable(tbl, filename, format = NULL)
}
\arguments{
\item{tbl}{Data frame to save}

\item{filename}{Output filename}

\item{format}{File format (csv, xlsx, json)}
}
\description{
Save Table to File
}
