% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_non_single_cat_cols}
\alias{select_non_single_cat_cols}
\title{Select non-single category columns}
\usage{
select_non_single_cat_cols(x)
}
\arguments{
\item{x}{Data frame.}
}
\value{
Data frame sin columnas de un único valor.
}
\description{
Elimina columnas que tienen un único valor (excepto indicador y count).
}
\keyword{internal}
