% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.R
\name{sib_available_grupos}
\alias{sib_available_grupos}
\title{Get available grupos}
\usage{
sib_available_grupos(tipo = NULL, con)
}
\arguments{
\item{tipo}{Tipo de grupo: "biologico" o "interes" (opcional).}

\item{con}{Conexión a la base de datos.}
}
\value{
Vector nombrado con slugs de grupos (nombres son las etiquetas).
}
\description{
Obtiene lista de grupos biológicos o de interés de conservación disponibles.
}
