% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.R
\name{sib_available_profile_types}
\alias{sib_available_profile_types}
\title{Get available profile types}
\usage{
sib_available_profile_types()
}
\value{
Vector con tipos de perfiles: "region", "territorio",
"grupo_biologico", "grupo_interes", "specie", "tematica".
}
\description{
Obtiene lista de tipos de perfiles disponibles.
}
