% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.R
\name{sib_available_subregions}
\alias{sib_available_subregions}
\title{Get available subregions}
\usage{
sib_available_subregions(region, con)
}
\arguments{
\item{region}{Slug de la región padre.}

\item{con}{Conexión a la base de datos.}
}
\value{
Vector con slugs de subregiones.
}
\description{
Obtiene lista de subregiones de una región padre.
}
