% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.R
\name{sib_available_tematicas}
\alias{sib_available_tematicas}
\title{Get available tematicas}
\usage{
sib_available_tematicas()
}
\value{
Vector nombrado con slugs de temáticas (nombres son las etiquetas
en español).
}
\description{
Obtiene lista de temáticas disponibles para filtrar datos.
}
