% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sib_region.R
\name{sib_calculate_region}
\alias{sib_calculate_region}
\title{Calculate region data}
\usage{
sib_calculate_region(region, vars = NULL, con = NULL)
}
\arguments{
\item{region}{Slug de la región.}

\item{vars}{Vector de nombres de variables a seleccionar (opcional).}

\item{con}{Conexión a la base de datos.}
}
\value{
Objeto \code{tbl} con datos de la región.
}
\description{
Calcula datos de una región combinando información de la tabla de regiones,
datos marinos y temáticas.
}
