% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{sib_chart_gt_table2}
\alias{sib_chart_gt_table2}
\title{Create gt table (version 2)}
\usage{
sib_chart_gt_table2(t, labels = NULL)
}
\arguments{
\item{t}{Data frame con dos columnas (especie y temática).}

\item{labels}{Vector con etiquetas para las columnas (opcional).}
}
\value{
Objeto gt con tabla.
}
\description{
Crea una tabla gt simple sin gráficos.
}
