% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{sib_chart_waffle}
\alias{sib_chart_waffle}
\title{Create waffle chart}
\usage{
sib_chart_waffle(d, path)
}
\arguments{
\item{d}{Data frame con datos.}

\item{path}{Ruta donde guardar el gráfico (opcional).}
}
\value{
Objeto ggplot con gráfico waffle.
}
\description{
Crea un gráfico waffle (cuadrícula) para mostrar proporciones.
}
