% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sib_merge_label.R
\name{sib_merge_especie_label}
\alias{sib_merge_especie_label}
\title{Merge especie labels into data}
\usage{
sib_merge_especie_label(x, con)
}
\arguments{
\item{x}{Data frame con columna \code{slug_especie}.}

\item{con}{Conexión a la base de datos.}
}
\value{
Data frame con columna \code{label} (nombre científico) agregada.
}
\description{
Agrega etiquetas de especies (nombres científicos) a un data frame.
}
