% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sib_merge_label.R
\name{sib_merge_grupo_label}
\alias{sib_merge_grupo_label}
\title{Merge grupo labels into data}
\usage{
sib_merge_grupo_label(d, slug, con)
}
\arguments{
\item{d}{Data frame con columna de slug de grupo.}

\item{slug}{Nombre de la columna con el slug de grupo ("slug" o
"slug_grupo").}

\item{con}{Conexión a la base de datos.}
}
\value{
Data frame con columna \code{label_grupo} agregada.
}
\description{
Agrega etiquetas de grupo biológico a un data frame usando el slug de grupo.
}
