% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sib_merge_label.R
\name{sib_merge_ind_label}
\alias{sib_merge_ind_label}
\title{Merge indicator labels into data}
\usage{
sib_merge_ind_label(d, replace = TRUE, con = con)
}
\arguments{
\item{d}{Data frame o vector de caracteres con nombres de indicadores.}

\item{replace}{Logical, reemplazar columna de indicador con etiqueta
(default: \code{TRUE}).}

\item{con}{Conexión a la base de datos.}
}
\value{
Data frame o vector con etiquetas de indicadores.
}
\description{
Agrega o reemplaza etiquetas de indicadores en un data frame o vector.
}
