% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sib_merge_label.R
\name{sib_merge_region_label}
\alias{sib_merge_region_label}
\title{Merge region labels into data}
\usage{
sib_merge_region_label(
  d,
  slug = "slug_region",
  label = "label_region",
  con = con
)
}
\arguments{
\item{d}{Data frame con columna de slug de región.}

\item{slug}{Nombre de la columna con el slug de región (default:
"slug_region").}

\item{label}{Nombre de la columna de etiqueta a crear (default:
"label_region").}

\item{con}{Conexión a la base de datos.}
}
\value{
Data frame con columna de etiqueta agregada.
}
\description{
Agrega etiquetas de región a un data frame usando el slug de región.
}
