% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sib_merge_label.R
\name{sib_merge_tematica_label}
\alias{sib_merge_tematica_label}
\title{Merge tematica labels into data}
\usage{
sib_merge_tematica_label(d, con)
}
\arguments{
\item{d}{Data frame con columna \code{slug_tematica}.}

\item{con}{Conexión a la base de datos.}
}
\value{
Data frame con columna \code{tematica_label} agregada.
}
\description{
Agrega etiquetas de temática a un data frame usando el slug de temática.
}
