% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parent.R
\name{sib_parent_region}
\alias{sib_parent_region}
\title{Get parent region}
\usage{
sib_parent_region(region, con)
}
\arguments{
\item{region}{Slug de la región.}

\item{con}{Conexión a la base de datos.}
}
\value{
Slug de la región padre.
}
\description{
Obtiene el slug de la región padre de una región dada, con casos especiales
para regiones naturales y áreas protegidas.
}
