% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sib_region.R
\name{sib_region_general}
\alias{sib_region_general}
\title{Get general region information}
\usage{
sib_region_general(region, con)
}
\arguments{
\item{region}{Slug de la región.}

\item{con}{Conexión a la base de datos.}
}
\value{
Lista con información general de la región.
}
\description{
Obtiene información general de una región incluyendo estadísticas,
descripción y créditos.
}
