% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sib_merge_label.R
\name{sib_region_labels}
\alias{sib_region_labels}
\title{Get region labels}
\usage{
sib_region_labels(con)
}
\arguments{
\item{con}{Conexión a la base de datos.}
}
\value{
Objeto \code{tbl} con columnas \code{slug} y \code{label}.
}
\description{
Obtiene tabla con slugs y etiquetas de regiones.
}
\keyword{internal}
