% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sib_region.R
\name{sib_region_marino}
\alias{sib_region_marino}
\title{Get marine regions data}
\usage{
sib_region_marino(con)
}
\arguments{
\item{con}{Conexión a la base de datos.}
}
\value{
Data frame con columnas \code{slug} y \code{marino} (logical).
}
\description{
Obtiene información sobre regiones marinas (departamentos, municipios y
Colombia).
}
\keyword{internal}
