% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgrupo.R
\name{sib_region_subgrupo}
\alias{sib_region_subgrupo}
\title{Get region subgrupo data}
\usage{
sib_region_subgrupo(region, grupo, con)
}
\arguments{
\item{region}{Slug de la región.}

\item{grupo}{Slug del grupo biológico padre.}

\item{con}{Conexión a la base de datos.}
}
\value{
Objeto \code{tbl} con datos de subgrupos.
}
\description{
Obtiene datos de subgrupos biológicos para una región y grupo padre.
}
