% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{sib_validate_available_regions}
\alias{sib_validate_available_regions}
\title{Validate region}
\usage{
sib_validate_available_regions(region)
}
\arguments{
\item{region}{Slug de la región a validar.}
}
\value{
Invisible, lanza error si la región no es válida.
}
\description{
Valida que una región esté disponible en la base de datos.
}
