% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{sib_validate_indicadores}
\alias{sib_validate_indicadores}
\title{Validate indicadores}
\usage{
sib_validate_indicadores(indicadores, con)
}
\arguments{
\item{indicadores}{Vector con nombres de indicadores a validar.}

\item{con}{Conexión a la base de datos.}
}
\value{
Invisible, lanza error si algún indicador no es válido.
}
\description{
Valida que todos los indicadores proporcionados estén disponibles en la base
de datos.
}
