% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{sib_validate_list_especies}
\alias{sib_validate_list_especies}
\title{Validate species list}
\usage{
sib_validate_list_especies(
  esps,
  region = NULL,
  grupo = NULL,
  tematica = NULL,
  validate = "warning"
)
}
\arguments{
\item{esps}{Data frame con lista de especies.}

\item{region}{Slug de la región.}

\item{grupo}{Slug del grupo biológico (opcional).}

\item{tematica}{Slug de la temática (opcional).}

\item{validate}{Tipo de validación: "warning" o "error" (default:
"warning").}
}
\value{
Invisible, lanza warning o error según \code{validate} si no coinciden.
}
\description{
Valida que el número de especies en una lista coincida con el número
esperado según los parámetros.
}
