% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{sib_validate_profile_type}
\alias{sib_validate_profile_type}
\title{Validate profile type}
\usage{
sib_validate_profile_type(type)
}
\arguments{
\item{type}{Tipo de perfil a validar.}
}
\value{
Invisible, lanza error si el tipo no es válido.
}
\description{
Valida que un tipo de perfil sea válido.
}
