% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{sibdata_aporte_region_especial}
\alias{sibdata_aporte_region_especial}
\title{Access aporte_region_especial table}
\usage{
sibdata_aporte_region_especial(con)
}
\arguments{
\item{con}{Conexión a la base de datos.}
}
\value{
Objeto \code{tbl} de dplyr para la tabla \code{aporte_region_especial}.
}
\description{
Access aporte_region_especial table
}
