% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{sibdata_banner_images}
\alias{sibdata_banner_images}
\title{Access banner_images table}
\usage{
sibdata_banner_images(con)
}
\arguments{
\item{con}{Conexión a la base de datos.}
}
\value{
Objeto \code{tbl} de dplyr para la tabla \code{banner_images}.
}
\description{
Access banner_images table
}
