% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{sibdata_dato_relevante}
\alias{sibdata_dato_relevante}
\title{Access dato_relevante table}
\usage{
sibdata_dato_relevante(con)
}
\arguments{
\item{con}{Conexión a la base de datos.}
}
\value{
Objeto \code{tbl} de dplyr para la tabla \code{dato_relevante}.
}
\description{
Access dato_relevante table
}
