% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{sibdata_departamento}
\alias{sibdata_departamento}
\title{Access departamento table}
\usage{
sibdata_departamento(con)
}
\arguments{
\item{con}{Conexión a la base de datos.}
}
\value{
Objeto \code{tbl} de dplyr para la tabla \code{departamento}.
}
\description{
Access departamento table
}
