% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{sibdata_especie}
\alias{sibdata_especie}
\title{Access especie table}
\usage{
sibdata_especie(con)
}
\arguments{
\item{con}{Conexión a la base de datos.}
}
\value{
Objeto \code{tbl} de dplyr para la tabla \code{especie}.
}
\description{
Access especie table
}
