% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{sibdata_especie_grupo}
\alias{sibdata_especie_grupo}
\title{Access especie_grupo table}
\usage{
sibdata_especie_grupo(con)
}
\arguments{
\item{con}{Conexión a la base de datos.}
}
\value{
Objeto \code{tbl} de dplyr para la tabla \code{especie_grupo}.
}
\description{
Access especie_grupo table
}
