% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{sibdata_especie_meta}
\alias{sibdata_especie_meta}
\title{Access especie_meta table}
\usage{
sibdata_especie_meta(con)
}
\arguments{
\item{con}{Conexión a la base de datos.}
}
\value{
Objeto \code{tbl} de dplyr para la tabla \code{especie_meta}.
}
\description{
Access especie_meta table
}
