% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{sibdata_especie_region}
\alias{sibdata_especie_region}
\title{Access especie_region table}
\usage{
sibdata_especie_region(con)
}
\arguments{
\item{con}{Conexión a la base de datos.}
}
\value{
Objeto \code{tbl} de dplyr para la tabla \code{especie_region}.
}
\description{
Access especie_region table
}
