% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{sibdata_especie_tematica}
\alias{sibdata_especie_tematica}
\title{Access especie_tematica table}
\usage{
sibdata_especie_tematica(con)
}
\arguments{
\item{con}{Conexión a la base de datos.}
}
\value{
Objeto \code{tbl} de dplyr para la tabla \code{especie_tematica}.
}
\description{
Access especie_tematica table
}
