% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{sibdata_estimada}
\alias{sibdata_estimada}
\title{Access estimada table}
\usage{
sibdata_estimada(con)
}
\arguments{
\item{con}{Conexión a la base de datos.}
}
\value{
Objeto \code{tbl} de dplyr para la tabla \code{estimada}.
}
\description{
Access estimada table
}
