% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{sibdata_gallery_images}
\alias{sibdata_gallery_images}
\title{Access gallery_images table}
\usage{
sibdata_gallery_images(con)
}
\arguments{
\item{con}{Conexión a la base de datos.}
}
\value{
Objeto \code{tbl} de dplyr para la tabla \code{gallery_images}.
}
\description{
Access gallery_images table
}
