% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{sibdata_glosario}
\alias{sibdata_glosario}
\title{Access glosario table}
\usage{
sibdata_glosario(con)
}
\arguments{
\item{con}{Conexión a la base de datos.}
}
\value{
Objeto \code{tbl} de dplyr para la tabla \code{glosario}.
}
\description{
Access glosario table
}
