% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{sibdata_grupo}
\alias{sibdata_grupo}
\title{Access grupo table}
\usage{
sibdata_grupo(con)
}
\arguments{
\item{con}{Conexión a la base de datos.}
}
\value{
Objeto \code{tbl} de dplyr para la tabla \code{grupo}.
}
\description{
Access grupo table
}
