% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{sibdata_ind_meta}
\alias{sibdata_ind_meta}
\title{Access ind_meta table}
\usage{
sibdata_ind_meta(con)
}
\arguments{
\item{con}{Conexión a la base de datos.}
}
\value{
Objeto \code{tbl} de dplyr para la tabla \code{ind_meta}.
}
\description{
Access ind_meta table
}
