% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores.R
\name{sibdata_indicadores}
\alias{sibdata_indicadores}
\title{Get indicators metadata}
\usage{
sibdata_indicadores(con)
}
\arguments{
\item{con}{Conexión a la base de datos.}
}
\value{
Data frame con metadatos de indicadores incluyendo columna
\code{subtematica}.
}
\description{
Obtiene metadatos de indicadores con subtemática calculada.
}
