% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{sibdata_municipio}
\alias{sibdata_municipio}
\title{Access municipio table}
\usage{
sibdata_municipio(con)
}
\arguments{
\item{con}{Conexión a la base de datos.}
}
\value{
Objeto \code{tbl} de dplyr para la tabla \code{municipio}.
}
\description{
Access municipio table
}
