% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{sibdata_patrocinador}
\alias{sibdata_patrocinador}
\title{Access patrocinador table}
\usage{
sibdata_patrocinador(con)
}
\arguments{
\item{con}{Conexión a la base de datos.}
}
\value{
Objeto \code{tbl} de dplyr para la tabla \code{patrocinador}.
}
\description{
Access patrocinador table
}
