% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{sibdata_preg_frecuentes}
\alias{sibdata_preg_frecuentes}
\title{Access preg_frecuentes table}
\usage{
sibdata_preg_frecuentes(con)
}
\arguments{
\item{con}{Conexión a la base de datos.}
}
\value{
Objeto \code{tbl} de dplyr para la tabla \code{preg_frecuentes}.
}
\description{
Access preg_frecuentes table
}
