% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{sibdata_publicador}
\alias{sibdata_publicador}
\title{Access publicador table}
\usage{
sibdata_publicador(con)
}
\arguments{
\item{con}{Conexión a la base de datos.}
}
\value{
Objeto \code{tbl} de dplyr para la tabla \code{publicador}.
}
\description{
Access publicador table
}
