% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{sibdata_ranking}
\alias{sibdata_ranking}
\title{Access ranking table}
\usage{
sibdata_ranking(con)
}
\arguments{
\item{con}{Conexión a la base de datos.}
}
\value{
Objeto \code{tbl} de dplyr para la tabla \code{ranking}.
}
\description{
Access ranking table
}
