% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{sibdata_referencia_estimada}
\alias{sibdata_referencia_estimada}
\title{Access referencia_estimada table}
\usage{
sibdata_referencia_estimada(con)
}
\arguments{
\item{con}{Conexión a la base de datos.}
}
\value{
Objeto \code{tbl} de dplyr para la tabla \code{referencia_estimada}.
}
\description{
Access referencia_estimada table
}
