% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{sibdata_referencias_home}
\alias{sibdata_referencias_home}
\title{Access referencias_home table}
\usage{
sibdata_referencias_home(con)
}
\arguments{
\item{con}{Conexión a la base de datos.}
}
\value{
Objeto \code{tbl} de dplyr para la tabla \code{referencias_home}.
}
\description{
Access referencias_home table
}
