% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{sibdata_region}
\alias{sibdata_region}
\title{Access region table}
\usage{
sibdata_region(con)
}
\arguments{
\item{con}{Conexión a la base de datos.}
}
\value{
Objeto \code{tbl} de dplyr para la tabla \code{region}.
}
\description{
Access region table
}
