% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{sibdata_region_grupo}
\alias{sibdata_region_grupo}
\title{Access region_grupo table}
\usage{
sibdata_region_grupo(con)
}
\arguments{
\item{con}{Conexión a la base de datos.}
}
\value{
Objeto \code{tbl} de dplyr para la tabla \code{region_grupo}.
}
\description{
Access region_grupo table
}
