% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{sibdata_region_patrocinador}
\alias{sibdata_region_patrocinador}
\title{Access region_patrocinador table}
\usage{
sibdata_region_patrocinador(con)
}
\arguments{
\item{con}{Conexión a la base de datos.}
}
\value{
Objeto \code{tbl} de dplyr para la tabla \code{region_patrocinador}.
}
\description{
Access region_patrocinador table
}
