% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{sibdata_region_publicador}
\alias{sibdata_region_publicador}
\title{Access region_publicador table}
\usage{
sibdata_region_publicador(con)
}
\arguments{
\item{con}{Conexión a la base de datos.}
}
\value{
Objeto \code{tbl} de dplyr para la tabla \code{region_publicador}.
}
\description{
Access region_publicador table
}
