% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{sibdata_region_tematica}
\alias{sibdata_region_tematica}
\title{Access region_tematica table}
\usage{
sibdata_region_tematica(con)
}
\arguments{
\item{con}{Conexión a la base de datos.}
}
\value{
Objeto \code{tbl} de dplyr para la tabla \code{region_tematica}.
}
\description{
Access region_tematica table
}
