% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{sibdata_tematica}
\alias{sibdata_tematica}
\title{Access tematica table}
\usage{
sibdata_tematica(con)
}
\arguments{
\item{con}{Conexión a la base de datos.}
}
\value{
Objeto \code{tbl} de dplyr para la tabla \code{tematica}.
}
\description{
Access tematica table
}
