% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tables.R
\name{sibdata_territorio}
\alias{sibdata_territorio}
\title{Access territorio table}
\usage{
sibdata_territorio(con)
}
\arguments{
\item{con}{Conexión a la base de datos.}
}
\value{
Objeto \code{tbl} de dplyr para la tabla \code{territorio}.
}
\description{
Access territorio table
}
