% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibdata.R
\name{sibdata_wide}
\alias{sibdata_wide}
\title{Query biodiversity data (wide format)}
\usage{
sibdata_wide(
  region = NULL,
  tipo = NULL,
  cobertura = NULL,
  tematica = NULL,
  grupo = NULL,
  indicador = NULL,
  subregiones = FALSE,
  with_parent = FALSE,
  con = NULL
)
}
\arguments{
\item{region}{Slug de la región.}

\item{tipo}{Tipo de dato (opcional).}

\item{cobertura}{Cobertura (opcional).}

\item{tematica}{Temática (opcional).}

\item{grupo}{Slug del grupo biológico (opcional).}

\item{indicador}{Indicador específico (opcional).}

\item{subregiones}{Logical, incluir subregiones.}

\item{with_parent}{Logical, incluir región padre.}

\item{con}{Conexión a la base de datos.}
}
\value{
Data frame en formato ancho.
}
\description{
Consulta datos de biodiversidad en formato ancho (una columna por
indicador).
}
