% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_clean}
\alias{str_clean}
\title{Clean string}
\usage{
str_clean(x)
}
\arguments{
\item{x}{Cadena de texto.}
}
\value{
Cadena limpia.
}
\description{
Limpia una cadena de texto: convierte a minúsculas, elimina acentos y
puntuación.
}
\keyword{internal}
