% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subregion.R
\name{subregion_tematica}
\alias{subregion_tematica}
\title{Get subregion tematica data}
\usage{
subregion_tematica(region, con)
}
\arguments{
\item{region}{Slug de la región padre.}

\item{con}{Conexión a la base de datos.}
}
\value{
Objeto \code{tbl} con datos de temática de subregiones.
}
\description{
Obtiene datos de temática para todas las subregiones de una región padre.
}
