% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sys_file_sibdata}
\alias{sys_file_sibdata}
\title{Get system file path for sibdata package}
\usage{
sys_file_sibdata(...)
}
\arguments{
\item{...}{Componentes de la ruta del archivo.}
}
\value{
Ruta completa del archivo del sistema.
}
\description{
Obtiene la ruta de archivos del sistema del paquete sibdata.
}
