% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tematica_list_col.R
\name{tematica_list_col}
\alias{tematica_list_col}
\title{Get tematica list for Colombia}
\usage{
tematica_list_col(region, con)
}
\arguments{
\item{region}{Slug de la región (debe ser "colombia").}

\item{con}{Conexión a la base de datos.}
}
\value{
Lista con datos de temáticas.
}
\description{
Genera lista completa de temáticas con indicadores y listas de especies para
Colombia (versión simplificada).
}
\keyword{internal}
