% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trim_punct}
\alias{trim_punct}
\title{Trim punctuation}
\usage{
trim_punct(x)
}
\arguments{
\item{x}{Cadena de texto.}
}
\value{
Cadena sin puntuación.
}
\description{
Elimina todos los caracteres de puntuación de una cadena.
}
\keyword{internal}
