% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_charts.R
\name{validate_chart_data}
\alias{validate_chart_data}
\title{Validate Chart Data}
\usage{
validate_chart_data(data, chart_type)
}
\arguments{
\item{data}{Data to validate}

\item{chart_type}{Type of chart}
}
\value{
TRUE if valid, FALSE otherwise
}
\description{
Validates that data has the required structure for charting
}
