% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotable_fun.R
\name{which_main_map}
\alias{which_main_map}
\title{Get main map name from map name}
\usage{
which_main_map(map_name)
}
\arguments{
\item{map_name}{Map name.}
}
\value{
Main map name.
}
\description{
Get main map name from map name
}
