% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parent.R
\name{with_parent_tematica}
\alias{with_parent_tematica}
\title{Get tematica data with parent region}
\usage{
with_parent_tematica(region, con = NULL)
}
\arguments{
\item{region}{Slug de la región.}

\item{con}{Conexión a la base de datos.}
}
\value{
Objeto \code{tbl} con datos de temática de la región y su padre.
}
\description{
Obtiene datos de temática incluyendo la región y su región padre.
}
\keyword{internal}
