% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{basic_map}
\alias{basic_map}
\title{Create basic map}
\usage{
basic_map(sf)
}
\arguments{
\item{sf}{Objeto \code{sf} con datos geoespaciales.}
}
\value{
Objeto Leaflet con mapa básico.
}
\description{
Crea un mapa básico sin datos asociados.
}
\keyword{internal}
