% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotable_fun.R
\name{default_icon_opts}
\alias{default_icon_opts}
\title{Default options for map icons}
\usage{
default_icon_opts()
}
\value{
List with default options for icons.
}
\description{
Default options for map icons
}
