% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotable_fun.R
\name{default_projections}
\alias{default_projections}
\title{Default cartographic projections}
\usage{
default_projections()
}
\value{
List with default projections.
}
\description{
Default cartographic projections
}
