% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_species_table.R
\name{exp_species_table_server}
\alias{exp_species_table_server}
\title{Species Table Server Module}
\usage{
exp_species_table_server(id, r, con)
}
\arguments{
\item{id}{Module ID}

\item{r}{Reactive values object}

\item{con}{Database connection}
}
\description{
Handles the server logic for species table including data fetching and modal display
}
