% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_visualization3.R
\name{exp_visualization3_server}
\alias{exp_visualization3_server}
\title{Visualization Server Module}
\usage{
exp_visualization3_server(id, r, con, loading_fns = NULL, debug = FALSE)
}
\arguments{
\item{id}{Module ID}

\item{r}{Reactive values object}

\item{con}{Database connection}

\item{debug}{Boolean to control console debug output}
}
\description{
Handles the server logic for visualization including chart rendering and data modals
}
