% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotable_fun.R
\name{gg_theme_nothing}
\alias{gg_theme_nothing}
\title{ggplot theme without visual elements}
\usage{
gg_theme_nothing(background_color = "transparent")
}
\arguments{
\item{background_color}{Background color of the plot (defaults to
"transparent").}
}
\value{
ggplot theme object (\code{theme}).
}
\description{
ggplot theme without visual elements
}
\keyword{internal}
