% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotable_fun.R
\name{gt_available_maps}
\alias{gt_available_maps}
\title{Get all available maps}
\usage{
gt_available_maps(con = NULL)
}
\arguments{
\item{con}{Connection (optional, defaults to \code{NULL}).}
}
\value{
Character vector of all available map names.
}
\description{
Get all available maps
}
