% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotable_fun.R
\name{gt_available_region_maps}
\alias{gt_available_region_maps}
\title{Get available region maps}
\usage{
gt_available_region_maps(con = NULL)
}
\arguments{
\item{con}{Connection (optional, defaults to \code{NULL}).}
}
\value{
Character vector of region map names.
}
\description{
Get available region maps
}
