% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotable_fun.R
\name{gt_discon}
\alias{gt_discon}
\title{Disconnect geotable connection}
\usage{
gt_discon(con)
}
\arguments{
\item{con}{Connection to the geotable database to disconnect.}
}
\value{
\code{NULL} (invisible).
}
\description{
Disconnect geotable connection
}
