% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotable_fun.R
\name{gt_sf}
\alias{gt_sf}
\title{Get Simple Features (sf) object from a map}
\usage{
gt_sf(map_name, con = NULL)
}
\arguments{
\item{map_name}{Name of the map to retrieve.}

\item{con}{Connection to the geotable database (optional, defaults to
\code{NULL}).}
}
\value{
\code{sf} object with the geographic data of the requested map.
}
\description{
Get Simple Features (sf) object from a map
}
\keyword{internal}
