% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_pages.R
\name{info_pages}
\alias{info_pages}
\title{Generate info pages data}
\usage{
info_pages(con)
}
\arguments{
\item{con}{Conexión a la base de datos.}
}
\value{
Lista con datos de publicadores, preguntas frecuentes, glosario y
tooltips.
}
\description{
Genera una lista con datos de páginas informativas.
}
