% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotable_fun.R
\name{rename_dotdot}
\alias{rename_dotdot}
\title{Rename columns with double dots}
\usage{
rename_dotdot(x)
}
\arguments{
\item{x}{\code{sf} object containing columns with names starting with "..".}
}
\value{
\code{sf} object with renamed columns.
}
\description{
Rename columns with double dots
}
