% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibdata.R
\name{select_indicator}
\alias{select_indicator}
\title{Select indicators based on filters}
\usage{
select_indicator(tipo = NULL, cobertura = NULL, tematica = NULL, con = NULL)
}
\arguments{
\item{tipo}{Tipo de dato (opcional).}

\item{cobertura}{Cobertura (opcional).}

\item{tematica}{Temática (opcional).}

\item{con}{Conexión a la base de datos.}
}
\value{
Vector con nombres de indicadores.
}
\description{
Selecciona indicadores disponibles según los filtros proporcionados.
}
\keyword{internal}
