% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.R
\name{sib_available_regions}
\alias{sib_available_regions}
\title{#' @export
sib_available_tables <- function(){
sibdata::available_tables
}}
\usage{
sib_available_regions(subtipo = NULL, departamento = NULL)
}
\description{
#' @export
sib_available_tables <- function(){
sibdata::available_tables
}
}
