% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subregion.R
\name{subregion_grupo}
\alias{subregion_grupo}
\title{Get subregion grupo data}
\usage{
subregion_grupo(region, grupo, con)
}
\arguments{
\item{region}{Slug de la región padre.}

\item{grupo}{Slug del grupo biológico.}

\item{con}{Conexión a la base de datos.}
}
\value{
Objeto \code{tbl} con datos de grupo de subregiones.
}
\description{
Obtiene datos de grupo biológico para todas las subregiones de una región
padre.
}
\keyword{internal}
