% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tematica_list.R
\name{tematica_list}
\alias{tematica_list}
\title{Get tematica list for region}
\usage{
tematica_list(region, con)
}
\arguments{
\item{region}{Slug de la región.}

\item{con}{Conexión a la base de datos.}
}
\value{
Lista con datos de temáticas (amenazadas, CITES, endémicas,
migratorias, exóticas).
}
\description{
Genera lista completa de temáticas con indicadores y listas de especies para
una región.
}
\keyword{internal}
